% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_list-class.R
\name{comp_list-class}
\alias{comp_list-class}
\title{Comparison between two lists of data frames}
\description{
Extension of class \linkS4class{comp_df} applied to comparisons of multiple data
frames.
}
