% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{credentials}
\alias{credentials}
\title{Prompt for credentials}
\usage{
credentials(user = "", password = "")
}
\arguments{
\item{user}{A character value with the name of the user. It is optional to
add it straight in the call of the function.}

\item{password}{A character value with the password. It is recommended to add
it manually in the prompt.}
}
\value{
A named vector with two elements, \strong{user} and \strong{password}.
}
\description{
Prompt for credentials
}
\details{
A wizard that asks for credentials (username and password).
This is to avoid writing passwords in the scripts.
The credentials are then passed to other functions that require them.
}
\author{
Miguel Alvarez
}
