% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_rows.R
\name{insert_rows}
\alias{insert_rows}
\alias{insert_rows,PostgreSQLConnection,data.frame-method}
\title{Insert new rows in database tables.}
\usage{
\S4method{insert_rows}{PostgreSQLConnection,data.frame}(x, y, name, eval = TRUE, ...)
}
\arguments{
\item{x}{A \linkS4class{PostgreSQLConnection} connecting to a database.}

\item{y}{A data frame including the reviewed version of the table.}

\item{name}{A character vector with the name of the schema and table in the
database. This function tests the existence of the table in the database
and returns an error message if not present.}

\item{eval}{A logical value, whether the resulting SQL commands should be
executed or not. This may be usefull if the target is retrieving SQL
scripts for further execution.}

\item{...}{Further arguments passed among methods. No yet in use.}
}
\value{
An invisible \linkS4class{sql} object.
}
\description{
New rows can be inserted from data frames into a Postgres table.
This function will first construct the sql commands to do the insert.
}
