% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sql.R
\name{read_sql}
\alias{read_sql}
\alias{write_sql}
\alias{write_sql.sql}
\alias{write_sql,sql-method}
\title{Read SQL scripts and split it in SQL statements}
\usage{
read_sql(file, ...)

write_sql(sql, ...)

\method{write_sql}{sql}(sql, file, ...)
}
\arguments{
\item{file}{A character value indicating the path to the sql file.}

\item{...}{Further arguments passed to \code{\link[=readLines]{readLines()}}.}

\item{sql}{A \linkS4class{sql} object containing queries.}
}
\value{
A \linkS4class{sql} object.
}
\description{
Read SQL scripts and split it in SQL statements
}
\details{
SQL scripts may mix SQL statements with comments. This function extract the
respective statements and split them into a list, where each element is a
statement.
}
\author{
Miguel Alvarez
}
