% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data.R
\name{reshape_updated}
\alias{reshape_updated}
\title{Reshape Updated Tables}
\usage{
reshape_updated(comp, key)
}
\arguments{
\item{comp}{A object of class \link[biblio:comp_df-class]{biblio::comp_df} resulted from
\code{\link[=compare_df]{compare_df()}}.}
}
\description{
Ease the update of modified tables usind SQL and loops
}
\keyword{internal}
