% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_rows.R
\name{update_rows}
\alias{update_rows}
\alias{update_rows,PostgreSQLConnection,data.frame-method}
\alias{update_rows,data.frame,data.frame-method}
\title{Row-wise update of database tables.}
\usage{
update_rows(x, y, ...)

\S4method{update_rows}{PostgreSQLConnection,data.frame}(x, y, name, key, eval = TRUE, ...)
}
\arguments{
\item{x}{A \link[RPostgreSQL:PostgreSQLConnection-class]{RPostgreSQL::PostgreSQLConnection} connecting to a
database.}

\item{y}{A data frame including identifiers and columns with values to be
updated.}

\item{...}{Addicional arguments passed among methods.}

\item{name}{A character vector with the name of the schema and table in the
database. This function tests the existence of the table in the database
and returns an error message if not present.}

\item{key}{A character vector with the columns of 'y' used as identifiers.
The occurrence of values or their combinations (if more then one
column set as key) will be tested for uniqueness and will retrieve an
error message if it fails.}

\item{eval}{A logical value, whether the resulting SQL commands should be
executed or not. This may be usefull if the target is retrieving SQL
scripts for further execution.}
}
\value{
An invisible \linkS4class{sql} object.
}
\description{
Updating database tables from data frames using a column as key value.
A SQL script will be created running row-wise updates.
}
