% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_columns.R
\name{add_columns}
\alias{add_columns}
\alias{add_columns.PostgreSQLConnection}
\alias{add_columns,PostgreSQLConnection-method}
\title{Add columns to an existing table in database}
\usage{
add_columns(conn, ...)

\method{add_columns}{PostgreSQLConnection}(conn, df, name, ...)
}
\arguments{
\item{conn}{A \linkS4class{PostgreSQLConnection} connection to a database.}

\item{df}{A data frame with definitions for new columns. The mandatory
columns are \strong{name}, \strong{type} (the type of new columns, including
constraints), and \strong{comment} (comments added to the columns).}

\item{name}{A character vector with the name of the schema and the table,
where the new columns have to be added.}
}
\description{
For convenience new columns can be listed in a data frame including type
and comments.
}
