% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_df.R
\name{compare_df}
\alias{compare_df}
\alias{compare_df,PostgreSQLConnection,data.frame,character-method}
\alias{compare_df,PqConnection,data.frame,character-method}
\alias{compare_df,list,list,list-method}
\title{Compare database tables with data frames}
\usage{
\S4method{compare_df}{PostgreSQLConnection,data.frame,character}(x, y, key, name, ...)

\S4method{compare_df}{PqConnection,data.frame,character}(x, y, key, name, ...)

\S4method{compare_df}{list,list,list}(x, y, key, ...)
}
\arguments{
\item{x}{A \link[RPostgreSQL:PostgreSQLConnection-class]{RPostgreSQL::PostgreSQLConnection} connecting to a database.}

\item{y}{A data frame including the reviewed version of the table.}

\item{key}{A character value indicating the name of the primary key in the
tables.}

\item{name}{A character vector with the name of the schema and table in the
database. This function tests the existence of the table in the database
and returns an error message if not present.}

\item{...}{Further arguments passed among methods. No yet in use.}
}
\value{
Either a \link[biblio:comp_df-class]{biblio::comp_df} or a \linkS4class{comp_list} object, depending on
the used method.
}
\description{
Database tables may be exported for editing as data frames. This method
allows for comparing editions done in data frame.
}
