% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_db.R
\name{connect_db}
\alias{connect_db}
\alias{disconnect_db}
\alias{disconnect_db.PostgreSQLConnection}
\alias{disconnect_db,PostgreSQLConnection-method}
\alias{disconnect_db.PqConnection}
\alias{disconnect_db,PqConnection-method}
\title{Connect to PostgreSQL database}
\usage{
connect_db(
  dbname,
  user,
  host = "localhost",
  port = "5432",
  pkg = "RPostgreSQL",
  ...
)

disconnect_db(conn, ...)

\method{disconnect_db}{PostgreSQLConnection}(conn, ...)

\method{disconnect_db}{PqConnection}(conn, ...)
}
\arguments{
\item{dbname, user}{Character values indicating the names of the database and
user. They are passed as service and username to \code{\link[keyring:key_get]{keyring::key_get()}},
respectively. Note that you need to set the respective password in
advance by using \code{\link[=credentials]{credentials()}}.}

\item{host, port}{Character values passed to \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}

\item{pkg}{A character value indicating the alternative package used to
establish the connection. At the moment only \code{RpostgreSQL} and
\code{RPostgres} are suitable.}

\item{...}{Further arguments passed to \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}

\item{conn}{A connection of class \link[RPostgreSQL:PostgreSQLConnection-class]{RPostgreSQL::PostgreSQLConnection}
or \link[RPostgres:PqConnection-class]{RPostgres::PqConnection}.}
}
\value{
A connection as \link[RPostgreSQL:PostgreSQLConnection-class]{RPostgreSQL::PostgreSQLConnection} or
\link[RPostgres:PqConnection-class]{RPostgres::PqConnection}, depending on the selected package.
}
\description{
It opens a prompt to insert connection details to a PostgreSQL
database.
Consequently \code{disconnect_db()} will be used to close the opened connection.
}
\author{
Miguel Alvarez \email{kamapu@posteo.com}
}
