% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_character.R
\name{do_character}
\alias{do_character}
\alias{do_character.character}
\alias{do_character,character-method}
\alias{do_character.numeric}
\alias{do_character,numeric-method}
\alias{do_character.factor}
\alias{do_character,factor-method}
\alias{do_character.logical}
\alias{do_character,logical-method}
\alias{do_character.Date}
\alias{do_character,Date-method}
\alias{do_character.data.frame}
\alias{do_character,data.frame-method}
\title{Convert verctors to characters for building queries}
\usage{
do_character(x, ...)

\method{do_character}{character}(x, ...)

\method{do_character}{numeric}(x, ...)

\method{do_character}{factor}(x, ...)

\method{do_character}{logical}(x, ...)

\method{do_character}{Date}(x, ...)

\method{do_character}{data.frame}(x, ...)
}
\arguments{
\item{x}{A vector or data frame to be converted.}

\item{...}{Further arguments passed among methods.}
}
\description{
To make it easier to build queries, variables and data frames are converted
to character values.
}
\examples{
## convert all columns of data set iris
iris_db <- do_character(iris)
str(iris_db)
}
