% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.R
\docType{methods}
\name{print-methods}
\alias{print-methods}
\alias{print.comp_list}
\title{Print methods}
\usage{
\method{print}{comp_list}(x, ...)
}
\arguments{
\item{x}{An object of class \linkS4class{comp_list}.}

\item{...}{Further arguments passed among methods.}
}
\description{
Printing outputs in diverse object classes.
}
