% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data.R
\name{update_data}
\alias{update_data}
\alias{update_data,PostgreSQLConnection,data.frame,character-method}
\title{Update database data by comparing with a reviewed data frame}
\usage{
\S4method{update_data}{PostgreSQLConnection,data.frame,character}(
  object,
  revision,
  key,
  name,
  eval = TRUE,
  add = FALSE,
  delete = FALSE,
  update = FALSE,
  ...
)
}
\arguments{
\item{object}{A connection as \link[RPostgreSQL:PostgreSQLConnection-class]{RPostgreSQL::PostgreSQLConnection}.}

\item{revision}{A data frame with the editions for 'object'.}

\item{key}{A character value indicating the name of the column used as
identifier for references.}

\item{name}{Character vector with two values indicating the name of the
schema and the name of the table in Postgres, respectively.}

\item{eval}{A logical value, whether the resulting SQL commands should be
executed or not. This may be usefull if the target is retrieving SQL
scripts for further execution.}

\item{add, delete, update}{Logical value indicating whether the respective
edition should be carried out in the database.}

\item{...}{Further arguments passed among methods. Not yet used.}
}
\value{
If all of add, delete and update are FALSE, this function returns an object
of class \link[biblio:comp_df-class]{biblio::comp_df}.
Otherwise, it returns an invisible \linkS4class{sql} object.
}
\description{
Edition can be done in a data frame imported from a database table. Those
editions can be then inserted in the original database table.

Note that you need to set the arguments \code{add}, \code{delete}, and \code{update} as
\code{TRUE} to carry out the respective actions.
If all \code{FALSE} as in the default, only a comparison will be done.
}
